+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Change Log
 * Added metadata_template.py module with goal of checking the user's metadata
   file with the latest HMS DBMI metadata template and giving a warning if the
   version appears to be out of date.
+* Added ability to print upload file info for check-submissionn.
 
 
 0.8.0